Refactor spiffe helper
