+++ tornjak-frontend/src/components/dashboard/agents-pie-chart.js
@@ -2,31 +2,29 @@ import React from 'react';
 import { connect } from 'react-redux';
 import Title from './title';
 import PieChart1 from "charts/PieChart";
-import SpiffeEntryInterface from '../spiffe-entry-interface'
-import SpiffeAgentInterface from '../spiffe-agent-interface'
+import SpiffeHelper from '../spiffe-helper'
 
 class AgentsPieChart extends React.Component {
   constructor(props) {
     super(props);
-    this.SpiffeEntryInterface = new SpiffeEntryInterface();
-    this.SpiffeAgentInterface = new SpiffeAgentInterface();
+    this.SpiffeHelper = new SpiffeHelper();
   }
 
   getChildEntries(agent, agentEntriesDict) {
-    var spiffeid = this.SpiffeAgentInterface.getAgentSpiffeid(agent);
+    var spiffeid = this.SpiffeHelper.getAgentSpiffeid(agent);
     var validIds = new Set([spiffeid]);
 
     // Also check for parent IDs associated with the agent
     let agentEntries = agentEntriesDict[spiffeid];
     if (agentEntries !== undefined) {
       for (let j=0; j < agentEntries.length; j++) {
-          validIds.add(this.SpiffeEntryInterface.getEntrySpiffeid(agentEntries[j]));
+          validIds.add(this.SpiffeHelper.getEntrySpiffeid(agentEntries[j]));
       }
     }
 
     if (typeof this.props.globalEntries.globalEntriesList !== 'undefined') {
       var check_id = this.props.globalEntries.globalEntriesList.filter(thisentry => {
-        return validIds.has(this.SpiffeEntryInterface.getEntryParentid(thisentry));
+        return validIds.has(this.SpiffeHelper.getEntryParentid(thisentry));
       });
     }
     if (typeof check_id === 'undefined') {
@@ -48,7 +46,7 @@ class AgentsPieChart extends React.Component {
         return [];
     }
 
-    let agentEntriesDict = this.SpiffeAgentInterface.getAgentsEntries(this.props.globalAgents.globalAgentsList, this.props.globalEntries.globalEntriesList)
+    let agentEntriesDict = this.SpiffeHelper.getAgentsEntries(this.props.globalAgents.globalAgentsList, this.props.globalEntries.globalEntriesList)
     var valueMapping = this.props.globalAgents.globalAgentsList.map(currentAgent => {
       return this.getChildEntries(currentAgent, agentEntriesDict);
     })
