--- tornjak-frontend/src/components/spiffe-entry-interface.js
@@ -1,91 +1,0 @@
-import { Component } from 'react';
-
-class SpiffeEntryInterface extends Component {
-  getMillisecondsFromEpoch() {
-    return new Date().getTime()
-  }
-
-  getAgentExpiryMillisecondsFromEpoch(entry) {
-    if (typeof entry !== 'undefined') {
-      return entry.x509svid_expires_at * 1000
-    }
-    return ""
-  }
-
-  getEntryExpiryMillisecondsFromEpoch(entry) {
-    if (typeof entry !== 'undefined' && typeof entry.expires_at !== 'undefined') {
-      return entry.expires_at * 1000
-    }
-    return ""
-  }
-
-  getEntryAdminFlag(entry) {
-    return typeof entry !== 'undefined' &&
-           typeof entry.admin !== 'undefined' &&
-           entry.admin;
-  }
-
-  // check if format strictly adhered to
-  checkSpiffeidValidity(trust_domain, path) {
-    if (typeof trust_domain === 'undefined' || typeof path === 'undefined') {
-      return false
-    } else if (trust_domain.length === 0 || path.length === 0) {
-        return false
-    } else if (trust_domain.charAt(0) === '/') {
-      return false
-    } else if (trust_domain.slice(-1) === "/" && path.charAt(0) !== "/") {
-      return true
-    }
-    return (trust_domain.slice(-1) !== "/" && path.charAt(0) === "/")
-  }
-
-  getAgentSpiffeid(entry) {
-    if (typeof entry !== 'undefined' && this.checkSpiffeidValidity(entry.id.trust_domain, entry.id.path)) {
-      return "spiffe://" + entry.id.trust_domain + entry.id.path
-    }
-    return ""
-  }
-
-  getEntrySpiffeid(entry) {
-    if (typeof entry !== 'undefined' && this.checkSpiffeidValidity(entry.spiffe_id.trust_domain, entry.spiffe_id.path)) {
-      return "spiffe://" + entry.spiffe_id.trust_domain + entry.spiffe_id.path
-    }
-    return ""
-  }
-
-  getEntryParentid(entry) {
-    if (typeof entry !== 'undefined' && this.checkSpiffeidValidity(entry.parent_id.trust_domain, entry.parent_id.path)) {
-      return "spiffe://" + entry.parent_id.trust_domain + entry.parent_id.path
-    }
-    return ""
-  }
-
-  getAgentStatusString(entry) {
-    if (typeof entry !== 'undefined') {
-      var banned = entry.banned
-      var status = "OK"
-      var expiry = this.getAgentExpiryMillisecondsFromEpoch(entry)
-      var currentTime = this.getMillisecondsFromEpoch()
-      if (banned) {
-        status = "Banned"
-      } else if (expiry > currentTime) {
-        status = "Attested"
-      }
-      return status
-    }
-    return ""
-  }
-
-  getAgentMetadata(spiffeid, metadataList) {
-    if (typeof metadataList !== 'undefined') {
-      var metadata = metadataList.filter(agent => (agent.spiffeid === spiffeid));
-      if (metadata.length !== 0) {
-        return metadata[0]
-      }
-    }
-    return {"plugin":"", "cluster":""}
-  }
-
-}
-
-export default SpiffeEntryInterface;
